Set extra_float_digits=3 to make reltuples precise
This change fixes the issue causing the query plans to deviate in outrows
slightly. 'reltuples' in pg_class has type real. By default Postgres rounds
real to a precision of 6 decimal places. The data is internally stored
as float and in C the precision of float is 7 digits. This leads to loss
of accuracy when migrating statistics. By setting the extra_float_digits
GUC to 3, we force Postgres to print accurate values upto 9 decimal places
which should be sufficient. The value of 3 is advised for historical reasons
in Postgres.