Test script for testing benchmark
To run this test we need a running instance of target and test Yugabyte
universes. These universes can be the same. The target universe should contain
dataset of the benchmark and ANALYZE should have been run on the tables.
The test will then run for each query in the test folder. For each
query, the test will export DDL, query plan and statistics from the
target database. It will then create a new database on the test
universe, create the empty tables and import statistics. It will export
the simulated query plan and diff against the query plan from the target
universe.